+++ src/views/MyVideosView/MyVideos.styles.ts
@@ -1,12 +1,27 @@
 import styled from '@emotion/styled'
-import { sizes } from '@/shared/theme'
-import { Text } from '@/shared/components'
-import { TOP_NAVBAR_HEIGHT } from '@/components/TopNavbar'
+import { colors, sizes } from '@/shared/theme'
+import { DismissibleMessage, Text } from '@/shared/components'
 
 export const ViewContainer = styled.div`
   padding-top: ${sizes(8)};
 `
-export const StyledText = styled(Text)`
-  /* Navbar Height padding so the text is not overlapped by Navbar when scrollIntoview */
-  /* padding-top: ${TOP_NAVBAR_HEIGHT}px; */
+export const StyledText = styled(Text)``
+
+export const TabsContainer = styled.div`
+  padding-top: ${sizes(8)};
+  margin-bottom: ${sizes(8)};
+  border-bottom: solid 1px ${colors.gray[800]};
+`
+
+type PaginationContainerProps = { extraPaddingBottom?: boolean }
+export const PaginationContainer = styled.div<PaginationContainerProps>`
+  padding-top: ${sizes(6)};
+  padding-bottom: ${({ extraPaddingBottom }) => (extraPaddingBottom ? sizes(24) : sizes(16))};
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`
+
+export const StyledDismissibleMessage = styled(DismissibleMessage)`
+  margin-bottom: ${sizes(8)};
 `
